Create a npm package that uses ES6 (with babel)
The package.json makes sure that ES6 code in src gets transpiled to ES5
code and saved in lib. npm scripts make sure that automatically happens
when the library is published.

The lib folder is not needed in the git repository and the src folder is
not needed in the npm package. I also ignored the test folder for
npm (to make sure I don't forget to add it when I add tests).

The prebuild event that deletes the lib folder is inspired by
stylelint (rimraf is platform independent).

I chose the folder name lib because the babel documentation uses that
name and a comment in the article referred to that.

See:
https://booker.codes/how-to-build-and-publish-es6-npm-modules-today-with-babel/
https://babeljs.io/docs/usage/cli/#compile-directories