Setup JavaScript linting with eslint
The --ignore-path .gitignore makes it possible to lint all JavaScript
files that are going to end up in the repository with a very easy
command.

The linting should be run during the testing process (realized with the
pretest script).

The eslint config is just my preferred way of writing JavaScript.