Setup Markdown linting with remark-lint
I chose the .js form of .remarkrc to get syntax highlighting when
editing the file.

The --quiet flag only outputs warnings and errors and --frail makes sure
that remark returns 1 when warnings occur (instead of return 1 only on
errors) which leads to failed builds on warnings.

The package npm-run-all makes running npm tasks in parallel possible. I
used run-p as the shorter version of npm-run-all --parallel.