Setup tests and code coverage
The tests should be saved in the tests directory with a filename that
coresponds to the file they are testing (with .test appended). The
.test.js makes it easier to distinguish them in the text editor and
automatically ignores the test files in the ncy coverage report.
See: https://github.com/bcoe/nyc#excluding-files

The glob that is passed to tape needs to be inside quotes or the unix
shell will expand the glob.
See: https://github.com/substack/tape#usage

Tape has an extra npm script to be able to run just the tests without
the pretest script.

The nyc option --all ensures that all files in the includes folder are
considered for the coverage report even when they are not required.
See: https://github.com/bcoe/nyc#include-reports-for-files-that-are-not-required

The require option for nyc is needed because the code and the tests are
written in ES6 and need to be transpiled to work.
See: https://github.com/bcoe/nyc#require-additional-modules

The dummy test is included to make setting up the travis and coveralls
configuration in later commits easier.