Use npm-watch to run tests when src/tests change
Added the `watch` script to make it easy to run npm-watch.

I run the tape command whenever a src or a test file changes (instead of
the test command) because linting errors can easily be shown in the
editor.

The quiet flag removes the script name from the npm-watch output which
is needed to get a correct layout for the coverage table.