Switch tests to jest
jest works with globals so they have to be added to the Eslint configuration.

I decided to move tests next to the files they test (jest picks up files that
are name *.test.js) to make them easier to locate.

The .test.js files need to be ignored when the application is build (they are
not needed to run and are shown as extra testcases otherwise).
https://github.com/skovhus/jest-codemods/blob/master/package.json
(scripts section)

jest comes with a coverage tool and a watch mode so nyc and npm-watch are no
longer needed.
https://github.com/nickmerwin/node-coveralls/issues/126#issuecomment-242963700

The jest configuration ensures that all src files are checked when the coverage
report is created (even when they are never required) and makes sure that the
compiled code is ignored.

babel-jest is needed to transform the code (including tests) before it is
executed. It just needs to be installed (will be used automatically). ES6
modules do not work without it.